make fsspec public
